+++ src/dotnet/APIView/APIViewWeb/Languages/TypeSpecLanguageService.cs
@@ -21,7 +21,7 @@ namespace APIViewWeb
 
         public override string Name { get; } = "TypeSpec";
         public override string [] Extensions { get; } = { ".tsp", ".cadl" };
-        public override string VersionString { get; } = "0";
+        public override string VersionString { get; } = "0.5.0";
         public override string ProcessName => throw new NotImplementedException();
 
         public TypeSpecLanguageService(IConfiguration configuration, TelemetryClient telemetryClient) : base(telemetryClient)
@@ -42,6 +42,11 @@ namespace APIViewWeb
         public override bool CanUpdate(string versionString)
         {
             return false;
+        }
+
+        public override bool CanConvert(string versionString)
+        {
+            return versionString != VersionString;
         }
 
         public override bool GeneratePipelineRunParams(APIRevisionGenerationPipelineParamModel param)