+++ tools/spec-gen-sdk/CHANGELOG.md
@@ -1,3 +1,20 @@
-## 2024-11-14 - 0.1.0
+# Release
+
+## 2025-01-14 - 0.1.3
+
+- Ensure the PrBranch variable is consistently set in all scenarios
+
+## 2025-01-07 - 0.1.2
+
+- Cleaned up dependencies
+- Set variables for pipeline scenarios
+- Added alias for CLI parameters
+
+## 2024-12-19 - 0.1.1
+
+- Added saveFilterLog function to save the filtered log
+- Introduced loggerWaitToFinish function to ensure log transports complete
+
+## 2024-12-17 - 0.1.0
 
 - Initial Release