+++ tools/tsp-client/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Release
 
+## 2025-01-10 - 0.15.2
+
+- Float `@autorest/openapi-to-typespec` version between `>=0.10.6 <1.0.0`.
+
+## 2025-01-09 - 0.15.1
+
+- Bumped `@autorest/openapi-to-typespec` version to `0.10.6`.
+
+## 2025-01-07 - 0.15.0
+
+- Support specifying an entrypoint file in tsp-location.yaml.
+- Ensure client.tsp selection over main.tsp in the entrypoint file search.
+
+## 2024-12-20 - 0.14.3
+
+- Bumped `@autorest/openapi-to-typespec` version to `0.10.5`.
+
 ## 2024-12-03 - 0.14.2
 
 - Bumped `@autorest/openapi-to-typespec` version to `0.10.4`.